add derivative method

+++ plot_breakpoint.py
@@ -1,10 +1,13 @@
-from algo.breakpoint import *
+from algo.breakpoint1 import *
+from algo.breakpoint2 import *
 import pandas as pd
 import imageio
 import shutil
 
 auxiliary_folder = './auxiliary'
 data_folder = './data'
+if_plot = False
+export_csv = True
 
 if __name__ == "__main__":
     data_train = np.load(f'{data_folder}/data_train.npy')
@@ -19,15 +22,28 @@ if __name__ == "__main__":
     # shape: 673, 187; num_samples, num_time_step
     train_labels = pd.read_csv('auxiliary/train_labels.csv', index_col="planet_id")
     file_names = []
+    
+    results = pd.DataFrame(columns=['planet_id', 'bound1', 'bound2', 'bound3', 'bound4'])
     for IDX in range(len(dataset)):
-    # for IDX in range(10):
+    # for IDX in range(630, 640):
         data = dataset[IDX]
-        find_and_plot_breakpoints(data, IDX, train_labels, verbose=True)
-        file_names.append(f'tmp/breakpoint_{IDX}.png')
+        bounds = find_derivative(data, IDX, train_labels, verbose=True, plot=if_plot)
+        if if_plot:
+            file_names.append(f'tmp/breakpoint_{IDX}.png')
+        planet_id = train_labels.index[IDX]
+        new_row = pd.DataFrame({
+        'planet_id': [planet_id],
+        'bound1': [bounds[0]],
+        'bound2': [bounds[1]],
+        'bound3': [bounds[2]],
+        'bound4': [bounds[3]],
+        })
+        results = pd.concat([results, new_row], ignore_index=True)
+    results.to_csv('breakpoints.csv', index=False)
     
-    with imageio.get_writer('breakpoints.mp4', mode='I', fps=1) as writer:
-        for filename in file_names:
-            image = imageio.imread(filename)
-            writer.append_data(image)
-    
-    shutil.rmtree('tmp')
+    if len(file_names) > 0:
+        with imageio.get_writer('breakpoints.mp4', mode='I', fps=1.5) as writer:
+            for filename in file_names:
+                image = imageio.imread(filename)
+                writer.append_data(image)
+        shutil.rmtree('tmp')
